Might as well use the library for all platforms
